Student_enrollment_system-main/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
mysql-client==0.0.1
mysqlclient==2.1.1
numpy==1.24.2
openpyxl==3.1.0
pandas==1.5.3
Pillow==9.4.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
reportlab==3.6.12
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
